refactor(www): Factor out modal to its own component (#21330)
* Factor out modal to its own component

* update TODO

* missing imports

* jsx pragma

* remove extraneous layout prop
